Extract history board from storm for puzzle history
